pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "deepradar"
version = "0.1.0"
description = "AI News Sniffer - Daily AI news and GitHub project aggregator"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.9,<4.0",
    "tenacity>=8.2,<9.0",
    "beautifulsoup4>=4.12,<5.0",
    "feedparser>=6.0,<7.0",
    "lxml>=5.0,<6.0",
    "anthropic>=0.40,<1.0",
    "pydantic>=2.5,<3.0",
    "pyyaml>=6.0,<7.0",
    "python-dateutil>=2.8,<3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4,<9.0",
    "pytest-asyncio>=0.23,<1.0",
    "aioresponses>=0.7,<1.0",
]
